[ADD] Add bin adaptation as argument to 1d hist
This extracts the bin adaptation policy into a separate class that must
be built and handed into the 1d histogram quantity if batch adaptation
is desired by the user. The functionality is very similar to the
``Quantity`` class, which is why the bin adaptation class directly
inherits from ``SingleStepQuantity``. Admittedly, this must be separated
more clearly at a later stage, but will likely not affect the policy API.